requirements.txt
streamlit>=1.39.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# Web search
ddgs>=9.0.0

# Environment & files
python-dotenv>=1.0.1
pypdf>=4.2.0
python-docx>=1.1.0
pyngrok>=7.2.0

# Extra LLM providers
langchain-google-genai>=1.0.0
langchain-deepseek>=0.1.0
langchain-xai>=0.1.0
